[REVIEW][MERGE] Display driver

+++ src/libs/plugins/prepare3d/main.cpp
@@ -316,7 +316,7 @@ void SimSourcePluginStateManager::processElement(char *element)
             el.value.string_value = value;
             el.length = strlen(value);
         }
-        else if (strncmp(type, "integer", sizeof(&type)) == 0) {
+        else if (strncmp(type, "int", sizeof(&type)) == 0) {
             el.type = CONFIG_INT;
             el.value.int_value = atoi(value);
             el.length = sizeof(int);
@@ -357,7 +357,7 @@ char *SimSourcePluginStateManager::getElementDataType(char identifier)
         return (char *)"float";
         break;
     case NUMBER_IDENTIFIER:
-        return (char *)"float";
+        return (char *)"int";
         break;
     case INDICATOR_IDENTIFIER:
         return (char *)"bool";
@@ -511,7 +511,7 @@ bool TCPClient::connect(std::string address, int port)
         else {
             // cast the h_addr_list to in_addr , since h_addr_list also has the ip
             // address in long format only
-            
+
             addr_list = (struct in_addr **)he->h_addr_list;
             for (int i = 0; addr_list[i] != NULL; i++) {
                 _server.sin_addr = *addr_list[i];
